get rid of messy object.h file, put it content in appropriate places
