+++ runtime/msgpack/adaptor/int.h
@@ -10,7 +10,6 @@
 #pragma once
 
 #include "runtime/msgpack/adaptor/adaptor_base.h"
-#include "runtime/msgpack/object.h"
 
 #include <limits>
